Techman-main/Techman-main/requirements.txt
# Tech Company Web Crawler and Job Openings Tracker
# Requirements File

# Web Crawling
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0

selenium==4.15.2
playwright==1.40.0
httpx==0.25.2
aiohttp==3.9.1
urllib3==2.1.0
fake-useragent==1.4.0
cloudscraper==1.2.66

# Database
mysql-connector-python==8.2.0
sqlalchemy==2.0.23
pymysql==1.1.0
redis==5.0.1
pymongo==4.6.1

# Message Queue & Task Management
celery==5.3.4
kombu==5.3.4
rq==1.15.1

# NLP & Text Processing
spacy==3.7.2
nltk==3.8.1
transformers==4.35.2
torch==2.2
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
regex==2023.10.3
phonenumbers==8.13.25

# AI & Interview Practice
openai==1.3.7
anthropic==0.18.1
speechrecognition==3.10.1
pyaudio==0.2.14
sounddevice==0.4.6
librosa==0.10.1
webrtcvad==2.0.10

# API & Server
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
python-dotenv==1.0.0
pyyaml==6.0.1

# Utilities
python-dateutil==2.8.2
 APScheduler==3.10.4
tenacity==8.2.3
tqdm==4.66.1
python-json-logger==2.0.7
structlog==24.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.23.3
pytest-cov==4.1.0
responses==0.24.1

# Monitoring
prometheus-client==0.19.0
psutil==5.9.7

# LinkedIn (if permitted)
linkedin-api==2.3.1

# Development
ipython==8.18.1
